feat(ios+shared): integrar trabajo de Jonathan — BLE wearable iOS, modelos shared, SOS
- BLEService.swift: escaneo y conexión BLE para wearable en iOS
- WearableView.swift + WearableViewModel.swift: pantalla de sincronización wearable
- AppNavigation.swift: rutas actualizadas
- AuthViewModel.swift, LoginView.swift, iosAppApp.swift: mejoras de autenticación iOS
- SharedModels.swift, Info.plist: actualización de modelos y permisos iOS
- SosManager shared + iOS: simplificación
- SleepData, VitalsData (shared): ajustes de modelos
- PatientDetailViewModel, RegisterViewModel: correcciones Android

+++ shared/src/commonMain/kotlin/mx/ita/vitalsense/data/model/SleepData.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class SleepData(
     val score: Int = 0,
-    val horas: Double = 0.0,
+    val horas: Float = 0f,
     val estado: String = "",
 )
